requirements.txt
fastapi
uvicorn
faster-whisper
edge-tts
requests
python-dotenv
asyncio
websockets
pydantic
aiohttp
